debugged mock

+++ software/src/enso/ixy_helpers.cpp
@@ -49,7 +49,7 @@
 namespace enso {
 
 #ifdef MOCK
-uint64_t virt_to_phys(void* virt) { return virt; }
+uint64_t virt_to_phys(void* virt) { return (uint64_t)virt; }
 #else
 uint64_t virt_to_phys(void* virt) {
   long page_size = sysconf(_SC_PAGESIZE);
